third_party/WebKit/LayoutTests/fast/dom/HTMLScriptElement/script-moving-from-non-contextdoc.html

<!DOCTYPE html>
<iframe src=resources/move-back-from-non-contextdoc.html></iframe>
<iframe src=resources/move-from-non-contextdoc-to-iframe.html></iframe>
<iframe src=resources/move-back-from-non-contextdoc.svg></iframe>
<script>
testRunner.waitUntilDone();
testRunner.dumpAsText();

// Don't use js-test.js because this bug is about <script> loading and
// it makes sense to keep this small and easy to debug when regress.
console.log("PASS unless crash.");

var count = document.querySelectorAll("iframe").length;

function done() {
    if (--count)
        return;
    testRunner.notifyDone();
}
</script>
