third_party/WebKit/LayoutTests/fast/dom/HTMLScriptElement/script-moving-from-non-contextdoc-expected.txt

CONSOLE MESSAGE: line 11: PASS unless crash.
  
